:bulb::recycle: 기프티콘 아이콘 클릭 시 기프티콘 이미지 보여주도록 구현 (#82)
- 기존에 WinView 내부에 fileprivate으로 존재하던 IconView를 GifticonIconView라는 파일을 생성하여 역할 대체함
